rename : AI 챗봇 관련 API 이름 변경 및 directory 위치 변경
