+++ src/main/java/com/example/classicjeans/controller/rest/ChatController.java
@@ -1,17 +1,18 @@
-package com.example.classicjeans.controller;
+package com.example.classicjeans.controller.rest;
 
 
 import com.example.classicjeans.entity.Users;
 import com.example.classicjeans.service.AlanSSEService;
 import com.example.classicjeans.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
 import java.util.concurrent.ConcurrentHashMap;
 
-@RequestMapping("/chat")
+@RequestMapping("api/chat")
 @RestController
 @RequiredArgsConstructor
 public class ChatController {
@@ -21,7 +22,7 @@ public class ChatController {
 
     // SSE 연결 설정
     @GetMapping("/stream")
-    public SseEmitter connect() {
+    public ResponseEntity<SseEmitter> connect() {
         SseEmitter emitter = new SseEmitter(0L); // 무제한 타임아웃
         Users user = SecurityUtil.getLoggedInUser();
         Long userId = user.getId();
@@ -34,6 +35,7 @@ public class ChatController {
             emitter.send(SseEmitter.event().name("userId").data(userId));
         } catch (IOException e) {
             emitters.remove(userId);
+            return ResponseEntity.status(500).body(null); // 오류 응답
         }
 
         emitter.onCompletion(() -> {
@@ -49,12 +51,12 @@ public class ChatController {
             System.out.println("SSE 연결 종료 (onError): " + userId);
         });
 
-        return emitter;
+        return ResponseEntity.ok(emitter);
     }
 
     // 메시지 전송 및 브로드캐스트
     @PostMapping("/send")
-    public void sendMessage(@RequestBody String content) {
+    public ResponseEntity<Void> sendMessage(@RequestBody String content) {
         emitters.forEach((id, emitter) -> {
             try {
                 // SSE를 통해 OpenAI API 스트리밍 응답 받기
@@ -64,14 +66,12 @@ public class ChatController {
                 emitters.remove(id);
             }
         });
+        return ResponseEntity.ok().build();
     }
 
     // SSE 연결 종료
     @PostMapping("/stream/close")
-    public void closeConnection(@RequestParam Long userId) {
-
-        // @RequestBody를 내가 Long으로 바꿨는데 그게 문제가 될까?
-
+    public ResponseEntity<Void> closeConnection(@RequestParam Long userId) {
         SseEmitter emitter = emitters.remove(userId);
         System.out.println("SSE 연결 종료 " + userId);
 
@@ -82,7 +82,9 @@ public class ChatController {
             } catch (Exception e) {
                 // 연결 종료 실패 처리
                 emitter.completeWithError(e);
+                return ResponseEntity.status(500).build();
             }
         }
+        return ResponseEntity.ok().build();
     }
 }
